requirements.txt
attrs==21.4.0
certifi==2021.10.8
colorama==0.4.4
commonmark==0.9.1
coverage==6.3.2
iniconfig==1.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.2
packaging==21.3
pip==21.2.4
pluggy==1.0.0
pprintpp==0.4.0
py==1.11.0
pyaml==20.4.0
Pygments==2.11.2
pyparsing==3.0.4
pytest==6.2.5
pytest-clarity==1.0.1
pytest-cov==3.0.0
pytest-sugar==0.9.4
PyYAML==6.0
rich==11.2.0
setuptools==58.0.4
six==1.16.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
wheel==0.37.1
